% Generated by roxygen2 (4.0.2): do not edit by hand
\name{b_brew}
\alias{b_brew}
\title{Use \code{\link{brew}} within a stage.}
\usage{
b_brew(filename)
}
\arguments{
\item{filename}{Path of the file to brew.}
}
\description{
This returns a function which can be passed as the \code{page} argument to
a Stage object. When the page is shown \code{\link{brew}} will be called
on \code{filename}. The variables \code{id}, \code{period}, \code{params}
and \code{error} will be available within the brew file. \code{params} is
a named list of HTML parameters. \code{error} is a character vector of errors,
e.g. from a form submission. Typically it will be \code{character(0)}.
}
\examples{
\dontrun{
text_stage(page=b_brew("mybrewfile.html"))
}
}
\seealso{
Other page creation: \code{\link{b_knit}}
}

