% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet.r
\name{bet}
\alias{bet}
\title{getThreshold}
\usage{
bet(df, minConc = 0, maxConc = NULL)
}
\arguments{
\item{df}{dataframe whose colnames are 'subject' 'log_concentration' and 'AFC'}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations). If NULL, calculated as the lowest concentration - the average of the mean difference between two consecutive concentrations}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations).If NULL, calculated as the lowest concentration + the average of the mean difference two consecutive concentrations}
}
\description{
bet returns the observed sensitivity thresholds of the subjects (=0 if the subject found the right sample for all concentrations, 1 if he/she found the right sample for all concentrations but the lightest one, etc.
}
\examples{
data(astringency)
bet(astringency)

}
