% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calculation}
\alias{calculation}
\title{Perform a calculation, either once or for each subject}
\usage{
calculation(fn, run = "once")
}
\arguments{
\item{fn}{a function which should take two arguments, \code{id} and
\code{period}.}

\item{run}{either "once" or "all".}
}
\value{
A Stage object of class Calculation
}
\description{
Perform a calculation, either once or for each subject
}
\details{
The function will be run when the first subject reaches this stage. If
\code{run} is \code{"all"} then it will be run again every time another
subject reaches the stage.
}
\examples{
expt <- experiment(N=4)
s1 <- text_stage(text="<html><body><form action=''>
     Enter a contribution:<input name='contrib' type='text'>
     <input type='submit' value='Next'></form></body></html>")

# go ahead individually:
add_stage(expt, s1, new_period(), times=2)

# wait for everyone:
add_stage(expt, s1, new_period("all"), times=2)

# players 1 and 2 wait for each other, so do 3 and 4:
add_stage(expt, s1, new_period(groups), times=2)
}
\seealso{
Other stages: \code{\link{NEXT}}, \code{\link{WAIT}},
  \code{\link{stage}}; \code{\link{period}};
  \code{\link{structured_stage}}; \code{\link{text_stage}}
}

