% Generated by roxygen2 (4.0.0): do not edit by hand
\name{checkpoint}
\alias{checkpoint}
\title{Make subjects wait for other subjects.}
\usage{
checkpoint(...)
}
\arguments{
\item{wait_for}{"all", "none", "ever" or a vector of length N}
}
\value{
An object of class CheckPoint
}
\description{
Make subjects wait for other subjects.
}
\details{
If \code{wait_for} is "all", then subjects must wait till all subjects in the
experiment have arrived. If \code{wait_for} is "none" then checkpoint does
nothing. If \code{wait_for} is "ever" then subjects wait forever (or until
moved on manually by the experimenter, using \code{\link{next_stage}}).

If \code{wait_for} is a vector, then it is assumed to represent
subject groups. So, if \code{wait_for} is \code{vec} then a subject with
ID \code{x} will wait here until all subjects with IDs
\code{which(vec==vec[x])} have arrived at this stage.

A typical use of this might be \code{mydf$groups[order(mydf$id),]}

If all relevant subjects are ready, the subject moves on.
}
\examples{
expt <- experiment(N=4)
groups <- c("A", "A", "B", "B")
s1 <- text_stage(page="<html><body><form action=''>
     <input type='submit' value='Next'></form></body></html>")

# wait for everyone:
add_stage(expt, period(), s1, checkpoint(), s1)

# players 1 and 2 wait for each other, so do 3 and 4:
add_stage(expt, period(), s1, checkpoint(c(1, 1, 2, 2)), s1)
}
\seealso{
Other stages: \code{\link{NEXT}}, \code{\link{WAIT}},
  \code{\link{stage}}; \code{\link{form_stage}};
  \code{\link{period}}; \code{\link{program}};
  \code{\link{text_stage}}; \code{\link{timed}}
}

