% Generated by roxygen2 (4.0.0): do not edit by hand
\name{experiment_data_frame}
\alias{experiment_data_frame}
\title{Create an appropriate data frame for an experiment}
\usage{
experiment_data_frame(experiment = NULL, N = NULL, periods = NULL)
}
\description{
This is a convenience function. It creates a data frame with \code{N * periods}
rows, with a column \code{id} varying from 1 to \code{N} and a column \code{period} varying
from 1 to \code{periods}. The data frame is sorted by ID, then by period.
\code{N} and \code{periods} can be specified manually, or you can just pass in
the \code{\link{experiment}} object and the function will guess for you.
}
\examples{
expt <- experiment(N=5)
s1 <- text_stage(text="<html><body>got here</body></html>")
add_stage(expt, s1, period(), times=5)
mydf <- experiment_data_frame(expt)
}
\seealso{
Other development tools: \code{\link{browser_test}};
  \code{\link{identify_seats}}
}

