% Generated by roxygen2 (4.0.0): do not edit by hand
\name{form_stage}
\alias{form_stage}
\title{Print out a form and store the subject's inputs in a data frame, after
checking for errors}
\usage{
form_stage(page, fields, titles = NULL, data_frame)
}
\arguments{
\item{page}{A character vector containing HTML, or a function to be called
with parameters \code{id, period, params, errors}.}

\item{fields}{A character vector of field names, or a list like
\code{list(field_name=check_function, ...)}}

\item{titles}{A list of field titles, like \code{list(field_name=title,...)}}

\item{data_frame}{The quoted string name of the data frame to be updated.
       This should exist in the global environment.}
}
\value{
An object of class FormStage
}
\description{
Print out a form and store the subject's inputs in a data frame, after
checking for errors
}
\details{
When the subject first arrives, the text or file \code{page} is
displayed. After the form is submitted, it is checked for errors, as follows:
each member of the list \code{fields} is called as a function, with arguments
\code{(field_title, value, id, period, params)}. Here

\itemize{
   \item \code{field_title} is the field name itself (the name of the
         corresponding list element) unless a list of titles was supplied.
   \item \code{value} is the value of the corresponding field in the form
         submitted by the user. This will be a length 1 character string.
   \item \code{id} and \code{period} are self-explanatory
   \item \code{params} is the full list of parameters
}

If the function returns \code{NULL} the field is OK. If the function
returns a character vector, this is treated as a vector of error messages.

If there are any error messages, \code{page} is redisplayed. If there
are no error messages, the data frame named in \code{data_frame} is updated:
in the row with \code{id==id && period==period}, the columns
named by \code{fields} get the values passed in by the user.

\code{\link{b_brew}} and similar functions can be used as arguments to
\code{page}.

\code{\link{is_whole_number}} and similar functions return functions suitable
for use in the fields list.
}
\examples{
mydf <- data.frame(id=rep(1:5, each=5), period=rep(1:5, times=5),
     username=NA_character_, password=NA_character_)

s1 <- form_stage(page=b_brew("myform.html"), data_frame="mydf",
 fields=list(
   username=length_between(8, 15),
   password=all_of(length_at_least(8), function(name, val, ...) {
     if (val=='password') return("Password should not be 'password'")
     if (grepl('&[A-Za-z0-9]$', val)) return("Password should contain at
           least one non-numeric character")
     return(NULL)
     })
))
}
\seealso{
Other stages: \code{\link{NEXT}}, \code{\link{WAIT}},
  \code{\link{stage}}; \code{\link{checkpoint}};
  \code{\link{period}}; \code{\link{program}};
  \code{\link{text_stage}}; \code{\link{timed}}
}

