% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{analyseSores}
\usage{
graph(
  df,
  subject = NULL,
  minConc = NULL,
  maxConc = NULL,
  y_add = NULL,
  ratio_ylim = 10
)
}
\arguments{
\item{df}{dataframe whose colnames are 'subject' 'log_concentration' and 'AFC'}

\item{subject}{if NULL (default), a graph of all subjects is produced. It could also contain the name of only one ore more subjects.}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations). If NULL, calculated as the lowest concentration - the average of the mean difference between two consecutive concentrations}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations).If NULL, calculated as the lowest concentration + the average of the mean difference two consecutive concentrations}

\item{y_add}{when representationAFC=='label', the label is displayed on the ordinate of the intensity + y_add}

\item{ratio_ylim}{default to 10. If y_add is NULL, y_add=(y_max-y_min)/ratio_ylim}
}
\description{
This function returns graphical results for threshold and intensity with concentrations as abscissa and intensities as ordinates. This is done by subject. In this case, the IBT and BET can be calculated and represented.
}
\details{
Returns the plots with the scores according to their concentrations
}
\examples{
data(astringency)
graph(astringency,subject="S002")
}
