% Generated by roxygen2 (4.0.0): do not edit by hand
\name{halt}
\alias{halt}
\title{Halt the experiment, stopping the server}
\usage{
halt(experiment, force = FALSE)
}
\arguments{
\item{experiment}{Object of class Experiment}

\item{force}{If TRUE, force the experiment to halt even if participants are
not finished.}
}
\value{
TRUE or FALSE, invisibly
}
\description{
Halt the experiment, stopping the server
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{nperiods}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{print_stages}}, \code{\link{session_name}};
  \code{\link{load_commands}}; \code{\link{next_stage}};
  \code{\link{pause}}; \code{\link{ready}};
  \code{\link{replay}}; \code{\link{restart}};
  \code{\link{start.Experiment}};
  \code{\link{trace_stage}}, \code{\link{untrace_stage}}
}

