% Generated by roxygen2 (4.0.0): do not edit by hand
\name{header}
\alias{footer}
\alias{header}
\title{Simple HTML header and footer}
\usage{
header(title = "Experiment", refresh = NA)

footer()
}
\description{
Convenience functions to print a simple HTML header and footer, optionally
with a title
}
\examples{
s1 <- text_stage(text=paste0(header(), "<b>Got here!</b>", footer()))
}
\seealso{
Other development tools: \code{\link{experiment_data_frame}};
  \code{\link{identify_seats}}; \code{\link{web_test}}
}

