% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibt.R
\name{ibt}
\alias{ibt}
\title{ibt}
\usage{
ibt(df, minConc = NULL, maxConc = NULL)
}
\arguments{
\item{df}{a dataframe containing at least 3 columns: 'subject','log_concentration' and 'intensity'}

\item{minConc}{a number indicating the minimal concentration (not tested, but used for threshold calculations). If NULL, calculated as the lowest concentration - the average of the mean difference between two consecutive concentrations}

\item{maxConc}{a number indicating the maximal concentration (not tested, but used for threshold calculations).If NULL, calculated as the lowest concentration + the average of the mean difference two consecutive concentrations}
}
\description{
A threshold is calculated as the start of the highest slope (cumulative while no negative slope)
}
\examples{
data(astringency)
ibt(astringency)
}
