% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_in.R
\name{load_users}
\alias{load_users}
\title{Load username listed in leagues meta.csv file}
\usage{
load_users(to_database)
}
\arguments{
\item{to_database}{path to database}
}
\description{
Load username listed in leagues meta.csv file
}
