% Generated by roxygen2 (4.0.2): do not edit by hand
\name{merge_subjects}
\alias{merge_subjects}
\title{Merge a data frame with information about experiment subjects}
\usage{
merge_subjects(experiment, data_frame)
}
\arguments{
\item{experiment}{an object of class Experiment}

\item{data_frame}{a data frame containing a column 'id'}
}
\value{
A new data frame produced by \code{\link{merge}} using the 'id'
column, resulting in new columns 'IP', 'client' and 'seat'. 'period', 'stage'
and 'status' will not be merged.
}
\description{
Merge a data frame with information about experiment subjects
}
\examples{
expt <- experiment(N=2, port=12345)
expt # on the command line, calls info()
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{nperiods}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{print_stages}}, \code{\link{session_name}};
  \code{\link{halt}}; \code{\link{load_commands}};
  \code{\link{next_stage}}; \code{\link{pause}};
  \code{\link{ready}}; \code{\link{replay}},
  \code{\link{rewind}}; \code{\link{restart}};
  \code{\link{start.Experiment}};
  \code{\link{trace_stage}}, \code{\link{untrace_stage}}
}

