% Generated by roxygen2 (4.0.0): do not edit by hand
\name{next_stage}
\alias{next_stage}
\title{Move some clients forward one stage}
\usage{
next_stage(experiment, subjid)
}
\arguments{
\item{experiment}{Object of class Experiment}

\item{subjid}{numeric vector of subject id(s) to move forward, or data
frame from subjects table}
}
\value{
TRUE or FALSE, invisibly
}
\description{
Manually moves one or more clients forward. This may break your experimental
design so use in emergencies only!
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{nperiods}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{print_stages}}, \code{\link{session_name}};
  \code{\link{halt}}; \code{\link{load_commands}};
  \code{\link{pause}}; \code{\link{ready}};
  \code{\link{replay}}; \code{\link{restart}};
  \code{\link{start.Experiment}};
  \code{\link{trace_stage}}, \code{\link{untrace_stage}}
}

