% Generated by roxygen2 (4.0.0): do not edit by hand
\name{on_ready}
\alias{on_ready}
\title{Set a function to be called whenever \code{ready} is called on an experiment.}
\usage{
on_ready(experiment, fn)
}
\arguments{
\item{epxt}{an Experiment object}

\item{fn}{a function, which should be callable no arguments}
}
\description{
Set a function to be called whenever \code{ready} is called on an experiment.
}
\details{
See \code{experiment} for more details.
}
\examples{
expt <- experiment(N=5, autostart=TRUE)
s1 <- stage(function(id, period, params) return("Got to s1!"))
add_stage(expt, s1, times=10)
on_ready(expt, function() {mydf <<- experiment_data_frame(expt)})
}

