% Generated by roxygen2 (4.0.0): do not edit by hand
\name{program}
\alias{program}
\title{Call a function, either once or for each subject}
\usage{
program(fn, run)
}
\arguments{
\item{fn}{a function which should take two arguments, \code{id} and
\code{period}.}

\item{run}{"first", "last", or "all".}
}
\value{
A Stage object of class Program
}
\description{
Call a function, either once or for each subject
}
\details{
If
\code{run} is \code{"all"} then the function \code{fn}will be run every time a
subject reaches the stage. If \code{run} is \code{"first"} then \code{fn}
will be run when the first subject reaches the stage. If \code{run} is
\code{"last"} then \code{fn} will be run when the last subject reaches
the stage.
}
\examples{
expt <- experiment(N=4)
s1 <- text_stage(text="<html><body><form action=''>
     Enter a contribution:<input name='contrib' type='text'>
     <input type='submit' value='Next'></form></body></html>")

s2 <- program
}
\seealso{
Other stages: \code{\link{NEXT}}, \code{\link{WAIT}},
  \code{\link{stage}}; \code{\link{checkpoint}};
  \code{\link{period}}; \code{\link{structured_stage}};
  \code{\link{text_stage}}
}

