% Generated by roxygen2 (4.0.0): do not edit by hand
\name{restart}
\alias{restart}
\title{Restart the experiment after pausing}
\usage{
restart(experiment)
}
\arguments{
\item{experiment}{Object of class Experiment}
}
\value{
TRUE or FALSE, invisibly
}
\description{
Experiments can be in status Stopped, Waiting, Started, or Paused.
\code{restart(experiment)} moves the experiment from Paused back to Started.
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{session_name}}; \code{\link{next_period}};
  \code{\link{pause}}; \code{\link{ready}};
  \code{\link{replay}}; \code{\link{start}}
}

