% Generated by roxygen2 (4.0.0): do not edit by hand
\name{start.Experiment}
\alias{start.Experiment}
\title{Start the experiment running.}
\usage{
\method{start}{Experiment}(experiment, force = FALSE)
}
\arguments{
\item{experiment}{Object of class Experiment}

\item{force}{Start the experiment even if there are fewer than N participants}
}
\value{
TRUE or FALSE, invisibly
}
\description{
Experiments can be in status Stopped, Waiting, Started, or Paused.
\code{start} moves the experiment from Waiting to Started.
If the experiment has autostart set, this will happen automatically
when N subjects have connected.
When the experiment starts, all subjects are moved to the first stage.
}
\details{
Note that \code{start} is an S3 generic to avoid clashes with \code{start} in
the stats package.
}
\examples{
start(expt)
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{nperiods}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{session_name}}; \code{\link{halt}};
  \code{\link{next_stage}}; \code{\link{pause}};
  \code{\link{ready}}; \code{\link{replay}};
  \code{\link{restart}}
}

