% Generated by roxygen2 (4.0.0): do not edit by hand
\name{trace_stage}
\alias{trace_stage}
\alias{untrace_stage}
\title{Trace one or more experiment stages}
\usage{
trace_stage(experiment, num, ...)

untrace_stage(experiment, num)
}
\arguments{
\item{experiment}{an object of class Experiment}

\item{num}{numbers of stages to trace}

\item{...}{arguments passed on to the \code{trace} method
       of the Stage object. See \code{\link{setRefClass}}.}
}
\description{
Trace one or more experiment stages
}
\examples{
\dontrun{
trace_stage(expt, s1, browser)
}
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{nperiods}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{print_stages}}, \code{\link{session_name}};
  \code{\link{halt}}; \code{\link{load_commands}};
  \code{\link{next_stage}}; \code{\link{pause}};
  \code{\link{ready}}; \code{\link{replay}};
  \code{\link{restart}}; \code{\link{start.Experiment}}
}

