% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_stage}
\alias{add_stage}
\title{Add one or more stages to an experiment}
\usage{
add_stage(experiment, ..., times, each, after)
}
\arguments{
\item{experiment}{an Experiment object}

\item{...}{one or more Stage objects, or functions}

\item{times}{how many times to repeat the sequence of stages in \code{...}.
All stages are repeated if this is a single number; if it is a
vector it gives how many times to repeat each stage.}

\item{each}{how many times to repeat each individual stage}

\item{after}{Add stages after how many stages (default: at the end)}
}
\description{
Add one or more stages to an experiment
}
\details{
If functions are passed in to \code{add_stage}, Stage objects will
automatically be created from them.
Stage objects are \link[=ReferenceClasses]{reference classes}. However,
when added to the experiment, they are copied. So, changing the
Stage after adding it to the experiment will not work.
}
\examples{
expt <- experiment(N=1, autostart=TRUE)
s1 <- stage(function(id, period, params) return("Got to s1!"))
# Or just define the function directly:
s2 <- function(id, period, params) return("Got to s2!")
add_stage(expt, s1, s2, times=2) # s1 s2 s1 s2
add_stage(expt, s1, s2, times=1:2) # s1 s2 s2
add_stage(expt, s1, s2, each=2) # s1 s1 s2 s2
info(expt)
}

