% Generated by roxygen2 (4.0.0): do not edit by hand
\name{b_brew}
\alias{b_brew}
\title{Use \code{\link{brew}} within a stage.}
\usage{
b_brew(filename)
}
\arguments{
\item{filename}{Path of the file to brew.}
}
\description{
This returns a function which can be passed as the \code{page} argument to
a Stage object. When the page is shown \code{\link{brew}} will be called
on \code{filename}. The variables \code{id}, \code{period} and \code{params}
will be available within the brew file.
}
\examples{
\dontrun{
text_stage(page=b_brew("mybrewfile.html"))
}
}

