% Generated by roxygen2 (4.0.0): do not edit by hand
\name{b_knit}
\alias{b_knit}
\title{Use \code{\link{knitr}} within a stage.}
\usage{
b_knit(filename)
}
\arguments{
\item{filename}{Path of the file to knit}
}
\description{
This returns a function which can be passed as the \code{page} argument to
a Stage object. When the page is shown \code{\link{knit}} will be called
on \code{filename}. The variables \code{id}, \code{period}, \code{params}
and \code{error} will be available within the knitr file. \code{params} is
a named list of HTML parameters. \code{error} is a character vector of errors,
e.g. from a form submission. Typically it will be \code{character(0)}.
}
\details{
By default the following options are set: \code{echo=FALSE, cache=FALSE,
warning=FALSE, message=FALSE} for \code{opts_chunk}; and
\code{out.format="html", progress=FALSE, upload.fun = image_uri} for
\code{opts_knit}. The last option means that images will be shown encoded as
"data URIs" in the document itself.
}
\examples{
\dontrun{
text_stage(page=b_knit("myknitfile.Rhtml"))
}
}
\seealso{
Other page creation: \code{\link{b_brew}}
}

