% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ready}
\alias{ready}
\title{Set the experiment up to receive participants.}
\usage{
ready(experiment)
}
\arguments{
\item{experiment}{Object of class Experiment}
}
\value{
TRUE or FALSE, invisibly
}
\description{
Experiments can be in status Stopped, Waiting, Started, or Paused.
\code{ready(experiment)} moves the experiment from Stopped to Waiting.
Clients can now connect and will be shown a waiting page.
}
\seealso{
Other command line functions: \code{\link{get_url}},
  \code{\link{info}}, \code{\link{map}},
  \code{\link{nperiods}},
  \code{\link{print,Experiment,ANY-method}},
  \code{\link{print,Experiment-method}},
  \code{\link{print_stages}}, \code{\link{session_name}};
  \code{\link{halt}}; \code{\link{load_commands}};
  \code{\link{next_stage}}; \code{\link{pause}};
  \code{\link{replay}}; \code{\link{restart}};
  \code{\link{start.Experiment}};
  \code{\link{trace_stage}}, \code{\link{untrace_stage}}
}

