% Generated by roxygen2 (4.0.2): do not edit by hand
\name{text_stage}
\alias{text_stage}
\title{Create a stage which does nothing but display HTML to the subject.}
\usage{
text_stage(page, wait = FALSE, name = "No name")
}
\arguments{
\item{page}{A character vector containing HTML, or a function to be called
with parameters \code{id, period, params}.}

\item{wait}{Wait to move on?}

\item{name}{optional name of the stage.}
}
\value{
An object of class TextStage. When called the first time, this will
        display the HTML in \code{file} or \code{text} to the participant.
        If \code{wait} is \code{FALSE} (the default), subsequent calls
        will return \code{NEXT}. If \code{wait} is \code{TRUE}, subsequent
        calls return the same page.
}
\description{
A text stage presents some HTML once to each subject.
}
\details{
It is always safe to call \code{next_stage} on a participant who
         is at a TextStage.

         Functions such as \code{\link{b_brew}} can be used as arguments to
         \code{page}.
}
\seealso{
Other stages: \code{\link{NEXT}}, \code{\link{WAIT}},
  \code{\link{stage}}; \code{\link{checkpoint}};
  \code{\link{form_stage}}; \code{\link{period}};
  \code{\link{program}}; \code{\link{timed}}
}

