% Generated by roxygen2 (4.0.0): do not edit by hand
\name{text_stage}
\alias{text_stage}
\title{Create a stage which does nothing but display HTML to the subject.}
\usage{
text_stage(...)
}
\arguments{
\item{file}{A filepath. If the file ends in ".brew" then it is passed to
\code{brew} for processing. Otherwise it is shown to the subject as-is.}

\item{text}{A character vector of HTML. Only one of \code{file} and
\code{text} should be passed.}
}
\value{
An object of class TextStage. When called the first time, this will
        display the HTML in \code{file} or \code{text} to the participant.
        Subsequent calls will return \code{NEXT}.
}
\description{
A text stage presents some HTML once to each subject.
}
\details{
It is always safe to call \code{next_stage} on a participant who
         is at a TextStage.
}
\seealso{
Other stages: \code{\link{NEXT}}, \code{\link{WAIT}},
  \code{\link{stage}}; \code{\link{period}};
  \code{\link{structured_stage}}
}

